Tie in Serde for change logging
