+++ winton_kafka_streams/kafka_streams.py
@@ -91,7 +91,7 @@ class KafkaStreams:
 
         self.consumer = None
 
-        self.stream_threads = [StreamThread(self.topology_builder, kafka_config, KafkaClientSupplier(self.kafka_config))
+        self.stream_threads = [StreamThread(self.topology_builder, self.kafka_config, KafkaClientSupplier(self.kafka_config))
                                for i in range(int(self.kafka_config.NUM_STREAM_THREADS))]
         for stream_thread in self.stream_threads:
             stream_thread.set_state_listener(self.on_thread_state_change)
